Move "INetworkProvider" and "IAbi" to controllers/interfaces.py.
